# static/openapi.yaml
openapi: 3.0.1
info:
  title: HANK.ai Medical AutoCoding Plugin
  description: A plugin that allows the user to submit medical documentation and returns a predicted claim to submit. User must enter an access token prior to using plugin. Responses may take up to 120 seconds.
  version: 'v1'
servers:
  - url: http://localhost:5003
  - url: https://chatgpt.api.hank.ai
# example user prompt:
# please code the following note for anesthesia services performed on the first monday of this month in columbia, sc. here's the nerve block note: Anesthesia Peripheral Block
#             Block Type: Adductor canal and pop sciatic nerve block
#             Patient Location during procedure: Post-op
#             Anesthesiologist: POKER, MASTER EDWARD
#             Injection Technique: Single-shots
#             Medications: Bupivacaine
#             Needle Gauge: 21 G
#             Needle Localization: Ultrasound guidance and anatomical landmarks
paths:
  /autocode/{username}/{token}:
    post:
      operationId: autocode
      summary: Autocode medical documents with cpts, icds, and modifiers
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The chatgpt username of the user.
      - in: path
        name: token
        schema:
          type: string
        required: true
        description: User must manually enter this token. This is their access token to the HANK.ai Autocoding plugin. If not yet provided by user, prompt the user to enter their token before posting request. User can contact support@hank.ai to request one.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/autocodeRequest'
components:
  schemas:
    autocodeRequest:
      type: object
      required:
      - autocodetext
      properties:
        autocodetext:
          type: string
          description: A summarized version of the medical document provided for to automated medical coding predictions.
          required: true
        dos:
          type: datetime
          description: The date of service for the encounter to be autocoded as YYYY-MM-DD
          required: true
        dob: 
          type: string
          description: The date of birth of the patient
          required: false
        # diagnosis_desc: 
        #   type: string
        #   description: The main diagnosis, as a short textual description, for the current encounter
        #   required: false
        # surgery_desc: 
        #   type: string
        #   description: The primary surgical procedure performed, as a short textual description, for the current encounter
        #   required: false

        zipcode:
          type: string
          description: The zip code where the medical encounter occurred
          required: false
        patient_age: 
          type: string
          description: The age of the patient
          required: false
        specialty:
          type: string
          description: The medical specialty providing care to the patient, which determines how the autocoder will code the encounter.
          enum: [anesthesia, surgical, radiology, emergencymedicine, eandm]
          required: true
        doc_type:
          type: string
          description: The type of medical document being submitted
          enum: [operative report, anesthesia procedure note, radiology report]
          required: true

    # SURGEON_OP_NOTE = "operative report"
    # ANESTHESIA_PREOP = "anes_preop"
    # ANESTHESIA_POSTOP = "anes_postop"
    # ANESTHESIA_INTRAOP = "anes_intraop"
    # ANESTHESIA_PROCEDURE_NOTE = "anesthesia procedure note"
    # H_AND_P = "history and physical"
    # PROGRESS_NOTE = "progress note"
    # RADIOLOGY = "radiology report"
    # OTHER = "medical documentation"